+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 * Moved `hdf5_availability()` from `Database` to `dataCAT.functions`.
 * Cleaned up the global namespace.
 * Added `libver="latest"` to `Database.hdf5`.
+* Updated the hdf5 logging system.
 
 
 0.3.0